+++ src/MobilityMap/MobilityMap.tsx
@@ -29,9 +29,27 @@ import Copyright from "../Copyright";
 import { I18nContext } from "../utils/hooks/useI18n";
 import StationsLayer from "../StationsLayer";
 import Station from "../Station";
-import { MobilityMapProps } from "./MobilityMapProps";
 import MobilityEvent from "../utils/MobilityEvent";
 import i18n from "../utils/i18n";
+
+export type MobilityMapProps = {
+  apikey?: string;
+  baselayer?: string;
+  center?: string;
+  geolocation?: string;
+  maxzoom?: string;
+  minzoom?: string;
+  mapsurl?: string;
+  mots?: string;
+  notification?: string;
+  notificationat?: string;
+  notificationurl?: string;
+  notificationbeforelayerid?: string;
+  realtime?: string;
+  realtimeurl?: string;
+  tenant?: string;
+  zoom?: string;
+};
 
 function MobilityMap({
   apikey = null,
