--- src/MobilityMap/MobilityMapProps.ts
@@ -1,20 +1,0 @@
-export type MobilityMapProps = {
-  apikey?: string;
-  baselayer?: string;
-  center?: string;
-  geolocation?: string;
-  maxzoom?: string;
-  minzoom?: string;
-  mapsurl?: string;
-  mots?: string;
-  notification?: string;
-  notificationat?: string;
-  notificationurl?: string;
-  notificationbeforelayerid?: string;
-  realtime?: string;
-  realtimeurl?: string;
-  tenant?: string;
-  zoom?: string;
-};
-
-export default MobilityMapProps;
